infrastructure/pyproject.toml
[project]
name = "infrastructure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.232.2",
    "constructs>=10.4.4",
    "ruff>=0.14.10",
]
